+++ lib/s_ink_button/src/s_ink_button.dart
@@ -405,6 +405,8 @@ class _SInkButtonState extends State<SInkButton> {
   @override
   Widget build(BuildContext context) {
     return MouseRegion(
+      cursor:
+          widget.isActive ? SystemMouseCursors.click : SystemMouseCursors.basic,
       onEnter: widget.isActive
           ? (_) {
               setState(() => _isHovered = true);